Use FQL for fetching event list and add rsvp_status.
New Graph API does not support anyway of fetching events that the user
has not RSVP'd as attending to. Need to refactor away from REST API
since it is going away, but use the FQL API instead. As a bonus,
we get access to the user's RSVP status in the exported event as well.